# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ResponseTRF:
    properties:
      amount:
        type: number
      targetUser:
        type: integer
      type:
        type: string
    type: object
...
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      fullname:
        type: string
      image:
        type: string
      phone:
        type: string
      pin:
        maxLength: 6
        minLength: 6
        type: string
      updatedAt:
        type: string
    type: object
  dto.UserSummaryDTO:
    properties:
      email:
        type: string
      fullname:
        type: string
      id:
        type: integer
      image:
        type: string
      phone:
        type: string
    type: object
  lib.PageInfo:
    properties:
      currentPage:
        type: integer
      nextPage:
        type: integer
      prevPage:
        type: integer
      totalData:
        type: integer
      totalPage:
        type: integer
    type: object
  lib.Response:
    properties:
      data: {}
      error: {}
      message:
        type: string
      pageInfo:
        $ref: '#/definitions/lib.PageInfo'
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: This is example for server P-Pay Aplication
  title: Backend P-Pay
  version: "1.0"
paths:
...
  /transfer/{id}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: transfer money
      parameters:
      - description: send to
        in: path
        name: id
        type: string
      - description: amount of money
        in: formData
        name: amount
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controllers.ResponseTRF'
      security:
      - ApiKeyAuth: []
      summary: Transfer
      tags:
      - Transaction
  /users:
    get:
      consumes:
      - application/json
      description: Get All Users
      parameters:
      - description: Search Users
        in: query
        name: search
        type: string
      - description: Page Users
        in: query
        name: page
        type: integer
      - description: Limit Users
        in: query
        name: limit
        type: integer
      - description: Sort Users
        in: query
        name: sort
        type: string
      - description: Order Users
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.UserSummaryDTO'
                  type: array
                pageInfo:
                  $ref: '#/definitions/lib.PageInfo'
              type: object
      summary: Users
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: User
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get Profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserSummaryDTO'
      security:
      - ApiKeyAuth: []
      tags:
      - Users
    patch:
      consumes:
      - multipart/form-data
      description: Update Movies
      parameters:
      - description: Update Full Name
        in: formData
        name: fullname
        type: string
      - description: Update Email
        in: formData
        name: email
        type: string
      - description: Update Password
        in: formData
        name: password
        type: string
      - description: Update Pin
        in: formData
        name: pin
        type: string
      - description: Update Phone
        in: formData
        name: phone
        type: string
      - description: Update Image
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/lib.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateUserRequest'
              type: object
      security:
      - ApiKeyAuth: []
      tags:
      - Users
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"